fix warnings on convertion from size_t to int

+++ ex_2/src/Board.cpp
@@ -5,7 +5,7 @@ Board::Board(std::string fileName)
 {
 	auto file = std::ifstream(fileName);
 
-	if (!file) 	        // if cant find file
+	if (!file) 	// if cant find file
 		exit(EXIT_FAILURE);
 
 	auto line = std::string();
@@ -17,26 +17,28 @@ Board::Board(std::string fileName)
 		{
 			if (line[col] == '%')
 			{
-				m_mouseFirstLoc.row = m_currBoard.size() - 1;
+				m_mouseFirstLoc.row = (int)(m_currBoard.size() - 1);
 				m_mouseFirstLoc.col = col;
 				m_currBoard[m_mouseFirstLoc.row][m_mouseFirstLoc.col] = ' ';
 			}
 			if (line[col] == '^')
 			{
-				m_catsFirstLocs.push_back(Location(col, m_currBoard.size() - 1));
+				m_catsFirstLocs.push_back(Location(col,(int)m_currBoard.size() - 1));
 				m_currBoard[m_currBoard.size() - 1][col] = ' ';  // Update the board
 			}
 			if (line[col] == '*')
-				m_cheeseLocations.emplace_back(Location(col, m_currBoard.size() - 1));
+				m_cheeseLocations.emplace_back(Location(col,(int)m_currBoard.size() - 1));
 		}
 	}
 }
 
+// returns current board size
 int Board::getBoardSize()
 {
-	return m_currBoard.size();
+	return (int)m_currBoard.size();
 }
 
+// prints current board
 void Board::printCurrBoard()
 {
 	for (auto row = 0; row < m_currBoard.size(); ++row)
@@ -49,27 +51,31 @@ void Board::printCurrBoard()
 	}
 }
 
+// returns a vector of first mouse location
 Location Board::getMouseLocation()
 {
 		return m_mouseFirstLoc;
 }
 
-// returns a vector of Cats Locations
+// returns a vector of Cats first Locations
 std::vector<Location> Board::getCatsLocations()
 {
 	return m_catsFirstLocs;
 }
 
+// returns a vector of cheese Locations on Board
 std::vector<Location> Board::getCheeseLocations()
 {
 	return m_cheeseLocations;
 }
 
+// clears a block on board
 void Board::clearBlock(Location position)
 {
 	m_currBoard[position.row][position.col] = ' ';
 }
 
+// remove the cheese from board and from cheese vector
 void Board::removeCheese(Location position)
 {
 	clearBlock(position);
@@ -84,6 +90,7 @@ void Board::removeCheese(Location position)
 	), m_cheeseLocations.end());
 }
 
+// checks if cheese left on board
 bool Board::isCheeseLeft()
 {
 	return !m_cheeseLocations.empty();
